Extract CallbackDrivenEntityMixin and device_info_for (no behavior change)
Stacked on the device-info-grouping fix: sensor.py and water_heater.py
reimplemented the same update_callback/async_update/
async_will_remove_from_hass triplet identically, and now build the
same device_info dict from separate copies. Extract both into
helpers.py so there's one definition of each instead of two, with
identical behavior.
