+++ custom_components/emeraldenergy/water_heater.py
@@ -25,6 +25,7 @@ from homeassistant.exceptions import HomeAssistantError
 from .const import (
     DOMAIN,
 )
+from .helpers import CallbackDrivenEntityMixin, device_info_for
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -93,7 +94,7 @@ async def async_setup_entry(
     return True
 
 
-class EmeraldWaterHeater(WaterHeaterEntity):
+class EmeraldWaterHeater(CallbackDrivenEntityMixin, WaterHeaterEntity):
     """Representation of a water heater."""
 
     def __init__(self, hass, emerald_hws_instance, hws_uuid, callback_dispatcher):
@@ -113,13 +114,9 @@ class EmeraldWaterHeater(WaterHeaterEntity):
         self._name = f"{self._brand} {self._serial_number}"
         # Same identifiers as sensor.py's device_info so both entities group
         # under one device instead of two.
-        self._attr_device_info = {
-            "identifiers": {(DOMAIN, hws_uuid)},
-            "name": self._name,
-            "manufacturer": self._brand,
-            "model": "Hot Water System",
-            "serial_number": self._serial_number,
-        }
+        self._attr_device_info = device_info_for(
+            hws_uuid, self._brand, self._serial_number
+        )
         self._current_temperature = status.get("last_state").get("temp_current")
         self._target_temperature = status.get("last_state").get("temp_set")
         self._running = emerald_hws_instance.isOn(hws_uuid)
@@ -240,22 +237,6 @@ class EmeraldWaterHeater(WaterHeaterEntity):
             _call_hws, "turn off", self._emerald_hws.turnOff, self._hws_uuid
         )
 
-    def update_callback(self):
-        """Schedules an update within HASS (called from the module's thread)."""
-        _LOGGER.info("emeraldhws: callback called")
-        if self.hass is None:
-            # The emerald_hws MQTT thread can fire callbacks before the entity
-            # is added to HASS (or after removal). schedule_update_ha_state is
-            # thread-safe, but with self.hass is None it would raise
-            # "'NoneType' object has no attribute 'create_task'".
-            _LOGGER.debug(
-                "Dropping callback for %s; hass not set (entity not added yet "
-                "or already removed)",
-                self._name,
-            )
-            return
-        self.schedule_update_ha_state(True)
-
     def update(self):
         """Update with values from HWS."""
         _LOGGER.info("emeraldhws: updating internal state from module")
@@ -268,13 +249,3 @@ class EmeraldWaterHeater(WaterHeaterEntity):
             self._current_mode = self._emerald_hws.currentMode(self._hws_uuid)
             self._is_heating = self._emerald_hws.isHeating(self._hws_uuid)
         return
-
-    async def async_update(self) -> None:
-        """Update the water heater state."""
-        await self._hass.async_add_executor_job(self.update)
-
-    async def async_will_remove_from_hass(self) -> None:
-        """Clean up when entity is removed from Home Assistant."""
-        # Unregister from callback dispatcher
-        self._callback_dispatcher.unregister_callback(self.update_callback)
-        await super().async_will_remove_from_hass()
